// src/app/interfaces/employerMember.ts
export class EmployerMember {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  employerTitle: string;
  employerID: number;
  emailVerified: number;
  status: number;
  profilePicture: string;
  createAt: Date;
  updateAt: Date;
  isDeleted: number;
  isUserDeleted: number
  permission: any;
}